Create ApiHelper class and use in test-options.ts
